==> 2. data_transformation.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt



# Definitions
pd.set_option('display.float_format', lambda x: '%.3f' % x)

njobs = -1
randomState = 0


%matplotlib inline

#data from previous step
data_acc = pd.read_csv('data_acc.csv', dtype={'LSOA_of_Accident_Location': str})
data_cas = pd.read_csv('data_cas.csv')
data_mak = pd.read_csv('data_mak.csv')

data_acc.columns

data_mak.head()

data_cas.head()

let us merge makemodel with casualties (this one in the left) over Accident_Index and Vehicles_Reference

#I will include a pedestrian factor, since pedestrian circunstances are related to my hypothesis.
data_cas['include_pedestrian'] = data_cas['Casualty_Class'].map(lambda x: 1 if x==3 else 0)

cas_mak = pd.merge(data_cas, data_mak, on= ['Accident_Index', 'Vehicle_Reference'], how = 'left')

now merge cas_mak in the left over Accident_Index

data = pd.merge(cas_mak,data_acc, on = 'Accident_Index', how = 'left')

unique_index = data['Accident_Index'].nunique()
unique_index

1. I need only one case per accident, I need to groupby per accident_Index
2. I want to test four hypotheses, oneway roads in rural areas, lack of pedestrian facilities,driver's age and  vehicle's age and IMD_decile : what varsiables are useful?



next step will be getting rid of useless variables 

data = data.set_index('Accident_Index')

data.columns

data = data.drop(['Vehicle_Reference', 'Casualty_Reference', 'Sex_of_Casualty', 'Age_of_Casualty', 
                  'Age_Band_of_Casualty',
                  'Casualty_Severity', 'Date', 'Car_Passenger', 'Bus_or_Coach_Passenger',
                  'Pedestrian_Road_Maintenance_Worker',
                 'Local_Authority_(District)', 'Local_Authority_(Highway)', 
                  'Accident_Severity', 'Towing_and_Articulation','Vehicle_Manoeuvre',
                  'Vehicle_Location-Restricted_Lane',
                  'Skidding_and_Overturning','Hit_Object_in_Carriageway','Vehicle_Leaving_Carriageway',
                  'Hit_Object_off_Carriageway', '1st_Point_of_Impact', 'Was_Vehicle_Left_Hand_Drive',
                  'Engine_Capacity_(CC)','Date','Propulsion_Code','Driver_Home_Area_Type', 'make',
                  'Time', 'Local_Authority_(District)', 'Local_Authority_(Highway)', 'Date_month',
                  'LSOA_of_Accident_Location',
    
                  'Casualty_Home_Area_Type'], axis = 1)

Vehicle_Reference and Casualty_Reference are now useless, Age_of_Casualty the same as Age_Band_of_Casualty, 'Date' is included in another way, Local_Authorities are identifiers, Casualty_Severity and Accident_Severity are both included in Class, Age_of_Vehicle is made in range mode. 


data.shape

Driver_Home_Area and  Casualty home area are considered in urban or rural area, Propulsion code, make, and engine capacity don't give useful  information for our purposes.
2ndRoad Class is an identifier

sex of casualty, age_band_of_casualty, car passenger, bus or coach passenger, pedestrian road maintenace worker, towing and articulation, vehicle maneouvre, vehicle location restricted line, junction location, skidding and overturning, 1st point of impact and hit circunstances are useless for our purposes

Casualty_Class now is included as being a pedestrian or not, the behavior of  pedestrian is not useful, junction location is related with kind of road and it is not informative 
Journey purpose, sex and Junction control are not useful 
Casualty type is not included because I am considering pedestrian circunstances as my first interest.

#get rid of something else
data = data.drop(['Vehicle_Type','Casualty_Class', 'Pedestrian_Location', 'Pedestrian_Movement',
                  'Casualty_Type',
                  'Class_Cas', 'Junction_Location', 'Vehicle_age_group',
                  'Journey_Purpose_of_Driver', 'Sex_of_Driver', 'Junction_Control'], axis = 1)


data.shape

data.columns

I will join vehicles table with casualty, variables interesting for our purposes are IMD deciles, age band of drivers and cars and include pedestrian, artificial index, 1 if there is pedestrian, 0 if not
Age of vehicle will be selected with max age, as well as include pedestrian

wea = data.groupby('Accident_Index', as_index=True).agg({'Casualty_IMD_Decile' : 'mean', 
                                                   'Age_Band_of_Driver' : 'mean', 'Driver_IMD_Decile': 'mean',
                                                  'Age_of_Vehicle' : 'max', 'include_pedestrian' : 'max'})


Some variables are useless for our purposes

#merge with accidents table

#data_acc = data_acc.set_index('Accident_Index')
new_data_acc = data_acc.drop([ 'Junction_Control','Local_Authority_(District)', 'Local_Authority_(Highway)', 
                  'Accident_Severity','Date',
                  'Time', 'Local_Authority_(District)', 'Local_Authority_(Highway)', 'Date_month',], axis = 1)

data_acc.columns

data = pd.merge(wea, new_data_acc, on='Accident_Index')

data = data.set_index('Accident_Index')

data.columns

data.shape

each column must be transformed into a more informative one

#into groups
data['Casualty_IMD_Decile'] = np.round(data['Casualty_IMD_Decile'], 0)
data['Age_Band_of_Driver'] = np.round(data['Age_Band_of_Driver'], 0)
data['Driver_IMD_Decile'] = np.round(data['Driver_IMD_Decile'], 0)


#group IMD, less levels will show  the trends in a better way
bins = [0,3,6,8,10]
groups = ['low','med_low','med_high','high']
data['Casualty_IMD_Group'] = pd.cut(data['Casualty_IMD_Decile'],bins,labels = groups)
data['Casualty_IMD_Group'].value_counts()
data = data.drop(['Casualty_IMD_Decile'], axis = 1)

#group here
bins = [0,3,6,8,10]
groups = ['low','med_low','med_high','high']
data['Driver_IMD_Group'] = pd.cut(data['Driver_IMD_Decile'],bins,labels = groups)
data['Driver_IMD_Group'].value_counts()
data = data.drop(['Driver_IMD_Decile'], axis = 1)

#group vehicle age
bins = [0,5,10,15,99]
groups = ['0-5','5-10','10-15','+15']
data['Vehicle_Age_Group'] = pd.cut(data['Age_of_Vehicle'],bins,labels = groups)
data['Vehicle_Age_Group'].value_counts()
data = data.drop(['Age_of_Vehicle'], axis = 1)

#group here
bins = [0,1,2,3,99]
groups = ['1','2','3','+4']
data['Number_Vehicles_Group'] = pd.cut(data['Number_of_Vehicles'],bins,labels = groups)
data['Number_Vehicles_Group'].value_counts()
data = data.drop(['Number_of_Vehicles'], axis = 1)

#group here
bins = [0,1,2,3,4,1000]
groups = ['1','2','3','4','+4']
data['Number_Casualties_Group'] = pd.cut(data['Number_of_Casualties'],bins,labels = groups)
data['Number_Casualties_Group'].value_counts()
data = data.drop(['Number_of_Casualties'], axis = 1)

#group if junction, if roundabout or other
data['Junction_Detail'].value_counts()

data['Junction_Group'] = data['Junction_Detail'].map(lambda x: 1 if x==0  else  2 if x==1 or x==2  else 3 )
data = data.drop(['Junction_Detail'], axis = 1)

#grpup if pedestrian have some facilities or not
data['Pedestrian_Control'] = data['Pedestrian_Crossing-Human_Control'].map(lambda x: 1 if x==0  else  2)
data = data.drop(['Pedestrian_Crossing-Human_Control'], axis = 1)
data['Pedestrian_Control'].value_counts()

#same as before
data['Pedestrian_PhisFac'] = data['Pedestrian_Crossing-Physical_Facilities'].map(lambda x: 1 if x==0  else  2)
data['Pedestrian_PhisFac'].value_counts()
data = data.drop(['Pedestrian_Crossing-Physical_Facilities'], axis = 1)

#I have grouped this considering avaiability of light
data['Active_Light'] = data['Light_Conditions'].map(lambda x: 0 if x==5 or x==6 or x==7  else  1 )
data = data.drop(['Light_Conditions'], axis = 1)


#good or bad
data['Weather'] = data['Weather_Conditions'].map(lambda x: 1 if x==1  else  0 )
data = data.drop(['Weather_Conditions'], axis = 1)
data.Weather.value_counts()

#good or bad
data['Road_Surf_Cond'] = data['Road_Surface_Conditions'].map(lambda x: 1 if x==1  else  0 )
data = data.drop(['Road_Surface_Conditions'], axis = 1)


#good or bad
data['Special_Conds'] = data['Special_Conditions_at_Site'].map(lambda x: 1 if x==0  else  0 )
data = data.drop(['Special_Conditions_at_Site'], axis = 1)

data.Special_Conds.value_counts()

#good or bad
data['Carriageway_Haz'] = data['Carriageway_Hazards'].map(lambda x: 1 if x==0  else  0 )
data = data.drop(['Carriageway_Hazards'], axis = 1)



data.Urban_or_Rural_Area.value_counts()

data.weekdays.value_counts()

#divide hour of the day in commute-non-commute 


data['Commute_hours'] = data['Time_hour'].map(lambda x: 1 if x in (7,9) or x in (15,17) else  0 )
#data = data.drop(['Carriageway_Hazards'], axis = 1)
data = data.drop(['Time_hour'], axis = 1)


#when no junction it is nan, in road number is 0, I will fillna with 0
data['2nd_Road_Class'] = data['2nd_Road_Class'].fillna(0)

data.shape

data.isna().sum()

10% rule: Do not impute data if more than 10% of column are NaNs. (Introduces artifial relationships)
Finally, I removed NaNs. Getting rid of them is the easiest solution but it has side effects, apart from lossing information we modify the distribution in classification problems, in this case, it means a proportion change between classes from 0.82 to 0.83.

data = data.dropna()
data.shape

data.to_csv('data_to_model.csv')

(data[data.Class == 0].count()[0]/data.shape[0])*100
