==> 3. Exploratory_Analysis.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns


from sklearn.model_selection import train_test_split    
from sklearn.metrics import accuracy_score,classification_report
from sklearn.preprocessing import LabelEncoder
from sklearn.ensemble import RandomForestClassifier
import statsmodels.api as sm


%matplotlib inline

data = pd.read_csv('data_to_model.csv')
data.head()

Lack of facilities is a reason for pedestrian accidents (81.2% of accidents involved pedestrians, only 2% of accidents involved pedestrian have Pedestrian Control, 32.8% have Physical Facilities

#proportion of serious accidents
data[data['include_pedestrian'] == 1]['Class'].value_counts()/data[data['include_pedestrian']==1].shape[0]

#include pedestrian, facilities or not?
include_pedestrian_t = (data[data['include_pedestrian']==1].Pedestrian_Control == 2).count()
include_ped_control = (data[data['include_pedestrian']==1].Pedestrian_Control == 2).sum()

proportion = include_ped_control/include_pedestrian_t
proportion

#include pedestrian, facilities or not?
include_pedestrian_t = (data[data['include_pedestrian']==1].Pedestrian_PhisFac == 2).count()
include_ped_control = (data[data['include_pedestrian']==1].Pedestrian_PhisFac == 2).sum()

proportion = include_ped_control/include_pedestrian_t
proportion

data.Pedestrian_PhisFac.value_counts()

data.Pedestrian_Control.value_counts()

data.Urban_or_Rural_Area.value_counts()

Singleway roads in rural areas are the most dangerous ones: 28% of rural accidents are serious vs 13% in urban roads. Bad road surface conditions are a casuse in 34% in rural roads versus 26% in urban roads.

#Proportion between serious in rural and urban areas
data[data['Urban_or_Rural_Area']==2]['Class'].sum()/data[data['Urban_or_Rural_Area']==2]['Class'].count()

data[data['Urban_or_Rural_Area']==1]['Class'].sum()/data[data['Urban_or_Rural_Area']==1]['Class'].count()

data.Road_Surf_Cond.value_counts()

#maintenace rural and urban areas
data[data['Urban_or_Rural_Area']==2]['Road_Surf_Cond'].sum()/data[data['Urban_or_Rural_Area']==2]['Road_Surf_Cond'].count()

data[data['Urban_or_Rural_Area']==1]['Road_Surf_Cond'].sum()/data[data['Urban_or_Rural_Area']==1]['Road_Surf_Cond'].count()

Serious accidents are more frequent in older vehicles. Vehicles less than 15 year-old are involved in 16% of the serious accidents while the figure for older vehicles grows up to 24.4%.


data.Vehicle_Age_Group.value_counts()

data['Class']==1
data[data['Class']==1].Vehicle_Age_Group.value_counts()/data.Vehicle_Age_Group.value_counts()

Serious accidents are more frequent with older drivers. Drivers in group 6 (26-35) are involved in serious accidents in 14% while the figure for groups 10 (56-65) and 11 (+75) grows to 22%.

data['Class']==1
data[data['Class']==1].Age_Band_of_Driver.value_counts()/data.Age_Band_of_Driver.value_counts()
