==> 4. hot_spots_map.ipynb <==
import pandas as pd
import numpy as np

I will consider a hot-spot when police have reported more than 25 accidents in a LSOA

data_p = pd.read_csv('data_to_model.csv')
data_p.head()


hot_spots = data_p.LSOA_of_Accident_Location.value_counts()>25

hot_spots['E01023190']

data_p['LSOA_of_Accident_Location'][9]

data_p['hot']= data_p.LSOA_of_Accident_Location.map(lambda x: hot_spots[x])
data_p['hot'].value_counts()

data_p = data_p[data_p['hot'] == True]

data_red = data_p[data_p['Class'] == 1]


#hot_spot = data_p.LSOA_of_Accident_Location == 'E01016486'

import folium
from folium import plugins

lat = data_p.Longitude.tolist()
lon = data_p.Latitude.tolist()
name = data_p.Accident_Index.tolist()
lat_red = data_red.Longitude.tolist()
lon_red = data_red.Latitude.tolist()
name_red = data_red.Accident_Index.tolist()

# Make a data frame with dots to show on the map
data = pd.DataFrame({
'lat':lat,
'lon':lon,
'name':name
})
data_red = pd.DataFrame({
'lat':lat_red,
'lon':lon_red,
'name':name_red
})
 
# Make an empty map
m = folium.Map(location=[51.5074, 0.1278],
                    zoom_start = 11 ,tiles="Stamen Terrain")
 
# I can add marker one by one on the map
for i in range(0,len(data)):
    folium.Marker([data.iloc[i]['lon'], data.iloc[i]['lat']],popup=data.iloc[i]['name'], icon=folium.Icon(color='blue')).add_to(m),  
for i in range(0,len(data_red)):
    folium.Marker([data_red.iloc[i]['lon'], data_red.iloc[i]['lat']],popup=data_red.iloc[i]['name'], icon=folium.Icon(color='red')).add_to(m),  


# Save it as html
m.save('hot_spots.html')
